+++ mental-health-assistance/package.json
@@ -4,9 +4,11 @@
   "private": true,
   "dependencies": {
     "bootstrap": "^5.3.3",
+    "chart.js": "^4.4.7",
     "cra-template": "1.2.0",
     "react": "^19.0.0",
     "react-bootstrap": "^2.10.7",
+    "react-chartjs-2": "^5.3.0",
     "react-dom": "^19.0.0",
     "react-icons": "^5.4.0",
     "react-router-dom": "^7.1.0",
