+++ mental-health-assistance/src/components/Footer.css
