+++ mental-health-assistance/src/components/Resources.js
@@ -61,6 +61,9 @@ const Resources = () => {
                             <div className="card-body">
                                 <h5 className="card-title">A 10 minute Guided Meditation for your Overactive Mind</h5>
                                 <p className="card-text">Listening to this 10-minute guided meditation can be a transformative experience for anyone struggling with an overactive mind.  </p>
+                                <div className="d-flex justify-content-end mr-3">
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
+                                </div>
                             </div>
                         </div>
                     </div>
@@ -70,6 +73,9 @@ const Resources = () => {
                             <div className="card-body">
                                 <h5 className="card-title">4 Small Habits To Improve Self Esteem</h5>
                                 <p className="card-text">Self esteem is defined as the degree to which qualities and characteristics inside one’s self-concept are perceived to be positive. In short, it is your opinion of yourself.</p>
+                                <div className="d-flex justify-content-end mr-3">
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
+                                </div>
                             </div>
                         </div>
                     </div>
@@ -79,6 +85,9 @@ const Resources = () => {
                             <div className="card-body">
                                 <h5 className="card-title">5 Ways to Deal with Anxiety</h5>
                                 <p className="card-text">Dr. Hansaji Yogendra explains how to deal with anxiety in 5 simple ways. Due to COVID-19, we have faced a lot of mental health issues in the society. This video will help you to deal with these issues.</p>
+                                <div className="d-flex justify-content-end mr-3">
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
+                                </div>
                             </div>
                         </div>
                     </div>
@@ -96,8 +105,9 @@ const Resources = () => {
                                 <h5 className="card-title">Importance of Regular Breaks and Mental Health Days</h5>
                                 <p className="card-text">
                                    
-                                        Encouraging employees to take regular breaks and mental health days is vital for maintaining a sustainable and productive work environment. Overworking and ignoring the need for rest can lead to burnout, decreased creativity, and poor decision-making.
-                                        <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info">Read Article</a>
+                                        Encouraging employees to take regular breaks and mental health days is vital for maintaining a sustainable and productive work environment. Overworking and ignoring the need for rest can lead to burnout.
+                                        <br/>
+                                        <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info mt-3">Read Article</a>
                                         <br />
                             </p>
             
@@ -111,9 +121,9 @@ const Resources = () => {
                     <div className="col-md-4">
                         <div className="card">
                             <div className="card-body">
-                                <h5 className="card-title">brain Eating Habits</h5>
+                                <h5 className="card-title">Brain Eating Habits</h5>
                                 <p className="card-text">Developing healthy eating habits is essential for maintaining overall well-being. A balanced diet provides the necessary nutrients for your body to function optimally, supports mental health, and reduces the risk of chronic diseases.</p>
-                                <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info">Read Article</a>
+                                <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info mt-3">Read Article</a>
                             </div>
                         </div>
                     </div>
@@ -122,7 +132,7 @@ const Resources = () => {
                             <div className="card-body">
                                 <h5 className="card-title">Exercise and Mental Health</h5>
                                 <p className="card-text">Physical activity has a profound impact on mental health. Regular exercise releases endorphins, which are natural mood lifters. It also helps reduce stress, anxiety, and symptoms of depression, while improving self-esteem and cognitive function.</p>
-                                <a href="https://example.com/article3" className="btn btn-info">Read Article</a>
+                                <a href="https://example.com/article3" className="btn btn-info mt-3">Read Article</a>
                             </div>
                         </div>
                     </div>
@@ -135,8 +145,13 @@ const Resources = () => {
                                 <h5 className="card-title">Importance of Regular Breaks and Mental Health Days</h5>
                                 <p className="card-text">
                                    
-                                        Encouraging employees to take regular breaks and mental health days is vital for maintaining a sustainable and productive work environment. Overworking and ignoring the need for rest can lead to burnout, decreased creativity, and poor decision-making.
+                                        Encouraging employees to take regular breaks and mental health days is vital for maintaining a sustainable and productive work environment. 
+                                        <div class="d-flex w-100 justify-content-between mt-3">
                                         <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info">Read Article</a>
+                                        <div className="d-flex mr-3">
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
+                                </div>
+                                        </div>
                                         <br />
                             </p>
             
@@ -150,9 +165,14 @@ const Resources = () => {
                     <div className="col-md-4">
                         <div className="card">
                             <div className="card-body">
-                                <h5 className="card-title">brain Eating Habits</h5>
+                                <h5 className="card-title">Brain Eating Habits</h5>
                                 <p className="card-text">Developing healthy eating habits is essential for maintaining overall well-being. A balanced diet provides the necessary nutrients for your body to function optimally, supports mental health, and reduces the risk of chronic diseases.</p>
-                                <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info">Read Article</a>
+                                <div class="d-flex w-100 justify-content-between mt-4">
+                                        <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info">Read Article</a>
+                                        <div className="d-flex mr-3">
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
+                                </div>
+                                        </div>
                             </div>
                         </div>
                     </div>
@@ -163,7 +183,12 @@ const Resources = () => {
                             <div className="card-body">
                                 <h5 className="card-title">Exercise and Mental Health</h5>
                                 <p className="card-text">Physical activity has a profound impact on mental health. Regular exercise releases endorphins, which are natural mood lifters. It also helps reduce stress, anxiety, and symptoms of depression, while improving self-esteem and cognitive function.</p>
-                                <a href="https://example.com/article3" className="btn btn-info">Read Article</a>
+                                <div class="d-flex w-100 justify-content-between">
+                                        <a href="https://www.youtube.com/watch?v=PBaFURjVrm0" className="btn btn-info">Read Article</a>
+                                        <div className="d-flex mr-3">
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
+                                </div>
+                                        </div>
                             </div>
                         </div>
                     </div>
@@ -228,8 +253,9 @@ const Resources = () => {
                                     <source src="https://example.com/podcast1.mp3" type="audio/mpeg" />
                                     Your browser does not support the audio element.
                                 </audio>
-                                <div className="d-flex justify-content-center">
+                                <div className="d-flex justify-content-between">
                                 <a href="https://example.com/podcast3" className="btn btn-primary mt-2">Listen to Podcast</a>
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
                                 </div>
                             </div>
                         </div>
@@ -243,8 +269,9 @@ const Resources = () => {
                                     <source src="https://example.com/podcast2.mp3" type="audio/mpeg" />
                                     Your browser does not support the audio element.
                                 </audio>
-                                <div className="d-flex justify-content-center">
+                                <div className="d-flex justify-content-between">
                                 <a href="https://example.com/podcast3" className="btn btn-primary mt-2">Listen to Podcast</a>
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
                                 </div>
                             </div>
                         </div>
@@ -258,8 +285,9 @@ const Resources = () => {
                                     <source src="https://example.com/podcast3.mp3" type="audio/mpeg" />
                                     Your browser does not support the audio element.
                                 </audio>
-                                <div className="d-flex justify-content-center">
+                                <div className="d-flex justify-content-between">
                                 <a href="https://example.com/podcast3" className="btn btn-primary mt-2">Listen to Podcast</a>
+                                <a href="#" className="btn btn-danger mt-2">Pay</a>
                                 </div>
                             </div>
                         </div>
