--- package.json
@@ -1,41 +1,0 @@
-{
-  "name": "mental-health-assistance",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "bootstrap": "^5.3.3",
-    "chart.js": "^4.4.7",
-    "cra-template": "1.2.0",
-    "react": "^19.0.0",
-    "react-bootstrap": "^2.10.7",
-    "react-chartjs-2": "^5.3.0",
-    "react-dom": "^19.0.0",
-    "react-icons": "^5.4.0",
-    "react-router-dom": "^7.1.0",
-    "react-scripts": "5.0.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
-}
