--- src/index.css
@@ -1,78 +1,0 @@
-body {
-  background-color: #f0f8ff;
-  color: #333;
-}
-
-.navbar-custom {
-  background-color: rgba(95, 158, 160, 0.8);
-  border-bottom: 2px solid #4f8a8b;
-}
-
-.navbar-brand {
-  font-size: 1.5rem;
-  font-weight: bold;
-  color: #fff;
-}
-
-.nav-link {
-  color: #fff;
-  margin-right: 15px;
-}
-
-.nav-link:hover {
-  color: #dcdcdc;
-}
-
-.footer {
-  background-color: rgba(224, 255, 255, 0.8);
-  padding: 20px;
-  text-align: center;
-}
-
-.card {
-  background-color: rgba(255, 255, 255, 0.9);
-  border: 1px solid #dcdcdc;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  margin-bottom: 20px;
-}
-
-.carousel-item {
-  display: flex;
-  justify-content: center;
-}
-
-.carousel-item .card {
-  width: 18rem;
-}
-
-.carousel-control-prev-icon,
-.carousel-control-next-icon {
-  background-color: #5f9ea0;
-  border-radius: 50%;
-}
-
-/* Section Background Colors */
-.section-videos {
-  background-color: #e6f7ff;
-  padding: 20px;
-  border-radius: 10px;
-}
-
-.section-articles {
-  background-color: #fff0f5;
-  padding: 20px;
-  border-radius: 10px;
-}
-
-.section-podcasts {
-  background-color: #f5f5dc;
-  padding: 20px;
-  border-radius: 10px;
-}
-
-.section-therapists {
-  background-color: #f0fff0;
-  padding: 20px;
-  border-radius: 10px;
-}
